requirements.txt
tensorflow-macos~=2.9.2
pandas~=1.4.3
numpy~=1.23.1
lemon~=1.1.0
python-dotenv~=0.19.0
keras~=2.9.0
backtesting~=0.3.3
matplotlib~=3.5.3
yfinance~=0.1.74
scikit-learn~=1.1.2